catcher-agent-worker/pyproject.toml
[project]
name = "catcher-agent-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.11.0",
    "temporalio>=1.19.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "temporal-lib-py>=1.8.4",
]
